Refs #372. refactored to move crystal classes to kernels/sample/crystal.h
